backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
django-phonenumber-field[phonenumbers]==6.3.0
djangorestframework==3.14.0
Pillow==8.2.0
environs[django]==9.3.2
requests==2.28.1
geopy==2.3.0
rollbar==1.0.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
